Add status page and merge private

+++ src/joy.c
@@ -1909,7 +1909,7 @@ glabel get_controller_buttons_held
 
 
 #ifdef NONMATCHING
-s32 get_controller_buttons_pressed(s32 arg0, s32 arg1) {
+s32 get_controller_buttons_pressed(s8 arg0, s32 arg1) {
     s32 temp_t7;
     void *temp_v1;
