+++ src/joy.h
@@ -4,6 +4,6 @@
 #include "ultra64.h"
 void *joyInitDebugNoticeList(void);
 s32 get_attached_controller_count(void);
-s32 get_controller_buttons_pressed(s32 arg0, s32 arg1);
+s32 get_controller_buttons_pressed(s8 arg0, s32 arg1);
 
 #endif
